:building_construction: refactored for consistency (WebHostTest* and WebHostFixture*)
