+++ src/Codebelt.Extensions.Xunit.Hosting.AspNetCore/IWebHostFixture.cs
@@ -6,8 +6,8 @@ namespace Codebelt.Extensions.Xunit.Hosting.AspNetCore
     /// <summary>
     /// Provides a way to use Microsoft Dependency Injection in unit tests tailored for ASP.NET Core.
     /// </summary>
-    /// <seealso cref="IHostFixture" />
-    public interface IAspNetCoreHostFixture : IHostFixture, IPipelineTest
+    /// <seealso cref="IGenericHostFixture" />
+    public interface IWebHostFixture : IGenericHostFixture, IPipelineTest
     {
         /// <summary>
         /// Gets or sets the delegate that configures the HTTP request pipeline.
